// frontend/Angular-workflow-diagram2-main/src/app/api/model/apiResponseChatIAResponse.ts
/**
 * Workflow Departamental API
 *
 * Contact: soporte@workflow.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ChatIAResponse } from './chatIAResponse';


export interface ApiResponseChatIAResponse { 
    exito?: boolean;
    mensaje?: string;
    datos?: ChatIAResponse;
    timestamp?: string;
    errores?: { [key: string]: Array<string>; };
}